Added: REST controller for machines to invalidate prescription.
