+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/Controllers/CasualUserAPIController.java
@@ -55,14 +55,4 @@ public class CasualUserAPIController {
         casualUsersService.reportBug(bug);
         return ResponseEntity.ok().build();
     }
-
-    @PutMapping("/invalidate")
-    public ResponseEntity<?> invalidatePrescription(@RequestParam int prescription_id){
-        boolean status = casualUsersService.invalidatePrescription(prescription_id);
-
-        if (!status)
-            return ResponseEntity.notFound().build();
-
-        return ResponseEntity.ok().build();
-    }
 }
